+++ README.md
@@ -9,7 +9,7 @@ Deployment: https://ucredit.me
 ## Getting Started
 
 1. "npm i"
-2. "npm run start" to run a local server on "localhost:3000" with nodemon listening for changes
+2. "npm start" to run a local server on "localhost:3000" with nodemon listening for changes
 3. "npm build" -> "npm start" to run an unlistened production-build server.
 4. The development branch has the newest development code. Master may not have the most updated code.
 5. All components asset files, redux files, etc. can be found in @src/components
